+++ lib/train-k8s-container.rb
@@ -1,17 +1,12 @@
 # frozen_string_literal: true
 
-require_relative "train/k8s/container/version"
-require_relative "train/k8s/container/platform"
-require_relative "train/k8s/container/connection"
-require_relative "train/k8s/container/transport"
-require_relative "train/k8s/container/kubectl_exec_client"
+libdir = File.dirname(__FILE__)
+$LOAD_PATH.unshift(libdir) unless $LOAD_PATH.include?(libdir)
 
-module Train
-  module K8s
-    module Container
-      class ConnectionError < StandardError
-      end
-      # Your code goes here...
-    end
-  end
-end
+require_relative 'train-k8s-container/version'
+require_relative 'train-k8s-container/errors'
+require_relative 'train-k8s-container/platform'
+require_relative 'train-k8s-container/retry_handler'
+require_relative 'train-k8s-container/transport'
+require_relative 'train-k8s-container/connection'
+require_relative 'train-k8s-container/kubectl_exec_client'